Remove RepoSync from URL
